Code refactor: Move db utilities to database, and other to utils
