+++ apps/website/app/api/supabase/content/route.ts
@@ -3,14 +3,14 @@ import type { PostgrestSingleResponse } from "@supabase/supabase-js";
 
 import { createClient } from "~/utils/supabase/server";
 import { getOrCreateEntity } from "~/utils/supabase/dbUtils";
+import { asPostgrestFailure } from "@repo/database/lib/contextFunctions";
 import {
   createApiResponse,
   handleRouteError,
   defaultOptionsHandler,
-  asPostgrestFailure,
 } from "~/utils/supabase/apiUtils";
 import { contentInputValidation } from "~/utils/supabase/validators";
-import { Tables, TablesInsert } from "@repo/database/types.gen.ts";
+import { Tables, TablesInsert } from "@repo/database/dbTypes";
 
 type ContentDataInput = TablesInsert<"Content">;
 type ContentRecord = Tables<"Content">;
