+++ apps/website/app/api/supabase/sync-task/[fn]/[target]/route.ts
@@ -1,11 +1,11 @@
 import { NextResponse, NextRequest } from "next/server";
 import type { PostgrestSingleResponse } from "@supabase/supabase-js";
 import { createClient } from "~/utils/supabase/server";
+import { asPostgrestFailure } from "@repo/database/lib/contextFunctions";
 import {
   createApiResponse,
   handleRouteError,
   defaultOptionsHandler,
-  asPostgrestFailure,
 } from "~/utils/supabase/apiUtils";
 
 type SyncTaskInfo = {
