+++ apps/website/app/utils/supabase/client.ts
@@ -1,5 +1,5 @@
 import { createClient as createSupabaseClient } from "@supabase/supabase-js";
-import { Database } from "@repo/database/types.gen.ts";
+import { Database } from "@repo/database/dbTypes";
 
 // Inspired by https://supabase.com/ui/docs/nextjs/password-based-auth
