+++ packages/database/package.json
@@ -5,9 +5,9 @@
   "license": "Apache-2.0",
   "type": "module",
   "exports": {
-    "./types.gen.ts": "./types.gen.ts",
-    "./types.gen": "./types.gen.ts",
-    "./dbDotEnv": "./dbDotEnv.ts"
+    "./lib/*": "./src/lib/*.ts",
+    "./dbDotEnv": "./src/dbDotEnv.ts",
+    "./dbTypes": "./src/dbTypes.ts"
   },
   "scripts": {
     "init": "supabase login",
@@ -19,13 +19,20 @@
     "build": "tsx scripts/build.ts && npm run genenv -- local",
     "test": "tsc && cucumber-js",
     "genenv": "tsx scripts/createEnv.ts",
-    "gentypes:production": "supabase start && supabase gen types typescript --project-id \"$SUPABASE_PROJECT_ID\" --schema public > types.gen.ts",
+    "gentypes:production": "supabase start && supabase gen types typescript --project-id \"$SUPABASE_PROJECT_ID\" --schema public > src/dbTypes.ts",
     "dbdiff": "supabase stop && supabase db diff",
     "dbdiff:save": "supabase stop && supabase db diff -f",
     "deploy": "tsx scripts/deploy.ts",
     "deploy:functions": "tsx scripts/lint.ts -f"
   },
+  "dependencies": {
+    "@repo/utils": "*",
+    "@supabase/auth-js": "2.71.1",
+    "@supabase/functions-js": "^2.4.5",
+    "@supabase/supabase-js": "2.55.0"
+  },
   "devDependencies": {
+    "@repo/typescript-config": "*",
     "@cucumber/cucumber": "^11.3.0",
     "@vercel/sdk": "^1.10.0",
     "prettier-plugin-gherkin": "^3.1.2",
