+++ packages/ui/package.json
@@ -35,8 +35,6 @@
     "typescript": "5.5.4"
   },
   "dependencies": {
-    "@supabase/functions-js": "^2.4.5",
-    "@supabase/supabase-js": "^2.52.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.468.0",
